Fix collision issue (mostly)
This fixes the collision issue with the seeds. Or well it mostly does. I have seen it still collide only slightly, but really not as bad as it was before, you can clearly tell how many of each color are in the holes now that's how good it is. Hope this helps!